requirements.txt
fastapi
uvicorn[standard]
google-genai
openai
reportlab
python-dotenv
jinja2
